+++ public/changelog.md
@@ -1,3 +1,7 @@
+# Update (12/10/2022) 9:04PM (Romania time)
+- Added Kronos
+- Added Zaphkiel. Paladin, Freyja
+
 # Startup (12/10/2022) 8:02PM (Romania Time) 19:03 (Austria Time)
 - Server Startup.
 - Changed FFA To Normal